web_app/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.2
pandas==2.1.4
python-dotenv==1.0.0
requests==2.31.0
gtfs-realtime-bindings==1.0.0
